// console/src/api/generated/models/link-feed-item-page.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.24.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { LinkFeedItem } from './link-feed-item';

/**
 *
 * @export
 * @interface LinkFeedItemPage
 */
export interface LinkFeedItemPage {
    /**
     *
     * @type {boolean}
     * @memberof LinkFeedItemPage
     */
    'hasNext'?: boolean;
    /**
     *
     * @type {Array<LinkFeedItem>}
     * @memberof LinkFeedItemPage
     */
    'items'?: Array<LinkFeedItem>;
    /**
     *
     * @type {string}
     * @memberof LinkFeedItemPage
     */
    'nextBeforeId'?: string;
    /**
     *
     * @type {string}
     * @memberof LinkFeedItemPage
     */
    'nextBeforePublishedAt'?: string;
}